Initial commit of generateReadmeFromHeader(s)
This adds a program that automatically generates
and updates a README from the header comments of
the files within the directory. It's functionality
is split in two: Most of the logic is handled by
an OS-agnostic C++ file, while a bash script
handles the linux/unix directory specific stuff

It also called to my attention that not all of
the files have header block comments. (oops!)

Future work: Soon, the program will be able to handle
a directory being passed in to it, so it can create
README's for any directory via a command line argument